simplify common/tier4_autoware_utils header
