+++ common/tier4_autoware_utils/CMakeLists.txt
@@ -8,10 +8,13 @@ find_package(Boost REQUIRED)
 
 ament_auto_add_library(tier4_autoware_utils SHARED
   src/tier4_autoware_utils.cpp
+  src/geometry/geometry.cpp
+  src/geometry/pose_deviation.cpp
   src/geometry/boost_polygon_utils.cpp
   src/math/sin_table.cpp
   src/math/trigonometry.cpp
   src/ros/msg_operation.cpp
+  src/ros/marker_helper.cpp
 )
 
 if(BUILD_TESTING)
