+++ common/tier4_autoware_utils/include/tier4_autoware_utils/geometry/boost_geometry.hpp
@@ -21,7 +21,6 @@
 
 #define EIGEN_MPL2_ONLY
 #include <Eigen/Core>
-#include <Eigen/Geometry>
 
 #include <geometry_msgs/msg/point.hpp>
