+++ common/tier4_autoware_utils/include/tier4_autoware_utils/geometry/boost_polygon_utils.hpp
@@ -22,9 +22,6 @@
 #include <autoware_auto_perception_msgs/msg/tracked_object.hpp>
 #include <geometry_msgs/msg/pose.hpp>
 
-#include <boost/geometry.hpp>
-#include <boost/optional.hpp>
-
 #include <vector>
 
 namespace tier4_autoware_utils
