+++ common/tier4_autoware_utils/include/tier4_autoware_utils/ros/marker_helper.hpp
@@ -15,7 +15,7 @@
 #ifndef TIER4_AUTOWARE_UTILS__ROS__MARKER_HELPER_HPP_
 #define TIER4_AUTOWARE_UTILS__ROS__MARKER_HELPER_HPP_
 
-#include <rclcpp/rclcpp.hpp>
+#include <rclcpp/time.hpp>
 
 #include <visualization_msgs/msg/marker_array.hpp>
 
@@ -64,56 +64,19 @@ inline std_msgs::msg::ColorRGBA createMarkerColor(float r, float g, float b, flo
   return color;
 }
 
-inline visualization_msgs::msg::Marker createDefaultMarker(
+visualization_msgs::msg::Marker createDefaultMarker(
   const std::string & frame_id, const rclcpp::Time & now, const std::string & ns, const int32_t id,
   const int32_t type, const geometry_msgs::msg::Vector3 & scale,
-  const std_msgs::msg::ColorRGBA & color)
-{
-  visualization_msgs::msg::Marker marker;
+  const std_msgs::msg::ColorRGBA & color);
 
-  marker.header.frame_id = frame_id;
-  marker.header.stamp = now;
-  marker.ns = ns;
-  marker.id = id;
-  marker.type = type;
-  marker.action = visualization_msgs::msg::Marker::ADD;
-  marker.lifetime = rclcpp::Duration::from_seconds(0.5);
+visualization_msgs::msg::Marker createDeletedDefaultMarker(
+  const rclcpp::Time & now, const std::string & ns, const int32_t id);
 
-  marker.pose.position = createMarkerPosition(0.0, 0.0, 0.0);
-  marker.pose.orientation = createMarkerOrientation(0.0, 0.0, 0.0, 1.0);
-  marker.scale = scale;
-  marker.color = color;
-  marker.frame_locked = true;
-
-  return marker;
-}
-
-inline visualization_msgs::msg::Marker createDeletedDefaultMarker(
-  const rclcpp::Time & now, const std::string & ns, const int32_t id)
-{
-  visualization_msgs::msg::Marker marker;
-
-  marker.header.stamp = now;
-  marker.ns = ns;
-  marker.id = id;
-  marker.action = visualization_msgs::msg::Marker::DELETE;
-
-  return marker;
-}
-
-inline void appendMarkerArray(
+void appendMarkerArray(
   const visualization_msgs::msg::MarkerArray & additional_marker_array,
   visualization_msgs::msg::MarkerArray * marker_array,
-  const boost::optional<rclcpp::Time> & current_time = {})
-{
-  for (const auto & marker : additional_marker_array.markers) {
-    marker_array->markers.push_back(marker);
+  const boost::optional<rclcpp::Time> & current_time = {});
 
-    if (current_time) {
-      marker_array->markers.back().header.stamp = current_time.get();
-    }
-  }
-}
 }  // namespace tier4_autoware_utils
 
 #endif  // TIER4_AUTOWARE_UTILS__ROS__MARKER_HELPER_HPP_
