+++ common/tier4_autoware_utils/src/geometry/geometry.cpp
@@ -1,0 +1,381 @@
+// Copyright 2023 TIER IV, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "tier4_autoware_utils/geometry/geometry.hpp"
+
+#include <Eigen/Geometry>
+
+#include <tf2/convert.h>
+namespace tf2
+{
+void fromMsg(const geometry_msgs::msg::PoseStamped & msg, tf2::Stamped<tf2::Transform> & out)
+{
+  out.stamp_ = tf2_ros::fromMsg(msg.header.stamp);
+  out.frame_id_ = msg.header.frame_id;
+  tf2::Transform tmp;
+  fromMsg(msg.pose, tmp);
+  out.setData(tmp);
+}
+}  // namespace tf2
+
+namespace tier4_autoware_utils
+{
+geometry_msgs::msg::Vector3 getRPY(const geometry_msgs::msg::Quaternion & quat)
+{
+  geometry_msgs::msg::Vector3 rpy;
+  tf2::Quaternion q(quat.x, quat.y, quat.z, quat.w);
+  tf2::Matrix3x3(q).getRPY(rpy.x, rpy.y, rpy.z);
+  return rpy;
+}
+
+geometry_msgs::msg::Vector3 getRPY(const geometry_msgs::msg::Pose & pose)
+{
+  return getRPY(pose.orientation);
+}
+
+geometry_msgs::msg::Vector3 getRPY(const geometry_msgs::msg::PoseStamped & pose)
+{
+  return getRPY(pose.pose);
+}
+
+geometry_msgs::msg::Vector3 getRPY(const geometry_msgs::msg::PoseWithCovarianceStamped & pose)
+{
+  return getRPY(pose.pose.pose);
+}
+
+geometry_msgs::msg::Quaternion createQuaternion(
+  const double x, const double y, const double z, const double w)
+{
+  geometry_msgs::msg::Quaternion q;
+  q.x = x;
+  q.y = y;
+  q.z = z;
+  q.w = w;
+  return q;
+}
+
+geometry_msgs::msg::Vector3 createTranslation(const double x, const double y, const double z)
+{
+  geometry_msgs::msg::Vector3 v;
+  v.x = x;
+  v.y = y;
+  v.z = z;
+  return v;
+}
+
+// Revival of tf::createQuaternionFromRPY
+// https://answers.ros.org/question/304397/recommended-way-to-construct-quaternion-from-rollpitchyaw-with-tf2/
+geometry_msgs::msg::Quaternion createQuaternionFromRPY(
+  const double roll, const double pitch, const double yaw)
+{
+  tf2::Quaternion q;
+  q.setRPY(roll, pitch, yaw);
+  return tf2::toMsg(q);
+}
+
+geometry_msgs::msg::Quaternion createQuaternionFromYaw(const double yaw)
+{
+  tf2::Quaternion q;
+  q.setRPY(0, 0, yaw);
+  return tf2::toMsg(q);
+}
+
+double calcElevationAngle(
+  const geometry_msgs::msg::Point & p_from, const geometry_msgs::msg::Point & p_to)
+{
+  const double dz = p_to.z - p_from.z;
+  const double dist_2d = calcDistance2d(p_from, p_to);
+  return std::atan2(dz, dist_2d);
+}
+
+/**
+ * @brief calculate azimuth angle of two points.
+ * @details This function returns the azimuth angle of the position of the two input points
+ *          with respect to the origin of their coordinate system.
+ *          If x and y of the two points are the same, the calculation result will be unstable.
+ * @param p_from source point
+ * @param p_to target point
+ * @return -pi < azimuth angle < pi.
+ */
+double calcAzimuthAngle(
+  const geometry_msgs::msg::Point & p_from, const geometry_msgs::msg::Point & p_to)
+{
+  const double dx = p_to.x - p_from.x;
+  const double dy = p_to.y - p_from.y;
+  return std::atan2(dy, dx);
+}
+
+geometry_msgs::msg::Pose transform2pose(const geometry_msgs::msg::Transform & transform)
+{
+  geometry_msgs::msg::Pose pose;
+  pose.position.x = transform.translation.x;
+  pose.position.y = transform.translation.y;
+  pose.position.z = transform.translation.z;
+  pose.orientation = transform.rotation;
+  return pose;
+}
+
+geometry_msgs::msg::PoseStamped transform2pose(
+  const geometry_msgs::msg::TransformStamped & transform)
+{
+  geometry_msgs::msg::PoseStamped pose;
+  pose.header = transform.header;
+  pose.pose = transform2pose(transform.transform);
+  return pose;
+}
+
+geometry_msgs::msg::Transform pose2transform(const geometry_msgs::msg::Pose & pose)
+{
+  geometry_msgs::msg::Transform transform;
+  transform.translation.x = pose.position.x;
+  transform.translation.y = pose.position.y;
+  transform.translation.z = pose.position.z;
+  transform.rotation = pose.orientation;
+  return transform;
+}
+
+geometry_msgs::msg::TransformStamped pose2transform(
+  const geometry_msgs::msg::PoseStamped & pose, const std::string & child_frame_id)
+{
+  geometry_msgs::msg::TransformStamped transform;
+  transform.header = pose.header;
+  transform.transform = pose2transform(pose.pose);
+  transform.child_frame_id = child_frame_id;
+  return transform;
+}
+
+Point3d transformPoint(const Point3d & point, const geometry_msgs::msg::Transform & transform)
+{
+  const auto & translation = transform.translation;
+  const auto & rotation = transform.rotation;
+
+  const Eigen::Translation3d T(translation.x, translation.y, translation.z);
+  const Eigen::Quaterniond R(rotation.w, rotation.x, rotation.y, rotation.z);
+
+  const Eigen::Vector3d transformed(T * R * point);
+
+  return Point3d{transformed.x(), transformed.y(), transformed.z()};
+}
+
+Point2d transformPoint(const Point2d & point, const geometry_msgs::msg::Transform & transform)
+{
+  Point3d point_3d{point.x(), point.y(), 0};
+  const auto transformed = transformPoint(point_3d, transform);
+  return Point2d{transformed.x(), transformed.y()};
+}
+
+Eigen::Vector3d transformPoint(const Eigen::Vector3d & point, const geometry_msgs::msg::Pose & pose)
+{
+  geometry_msgs::msg::Transform transform;
+  transform.translation.x = pose.position.x;
+  transform.translation.y = pose.position.y;
+  transform.translation.z = pose.position.z;
+  transform.rotation = pose.orientation;
+
+  Point3d p = transformPoint(Point3d(point.x(), point.y(), point.z()), transform);
+  return Eigen::Vector3d(p.x(), p.y(), p.z());
+}
+
+geometry_msgs::msg::Point transformPoint(
+  const geometry_msgs::msg::Point & point, const geometry_msgs::msg::Pose & pose)
+{
+  const Eigen::Vector3d vec = Eigen::Vector3d(point.x, point.y, point.z);
+  auto transformed_vec = transformPoint(vec, pose);
+
+  geometry_msgs::msg::Point transformed_point;
+  transformed_point.x = transformed_vec.x();
+  transformed_point.y = transformed_vec.y();
+  transformed_point.z = transformed_vec.z();
+  return transformed_point;
+}
+
+geometry_msgs::msg::Point32 transformPoint(
+  const geometry_msgs::msg::Point32 & point32, const geometry_msgs::msg::Pose & pose)
+{
+  const auto point =
+    geometry_msgs::build<geometry_msgs::msg::Point>().x(point32.x).y(point32.y).z(point32.z);
+  const auto transformed_point = tier4_autoware_utils::transformPoint(point, pose);
+  return geometry_msgs::build<geometry_msgs::msg::Point32>()
+    .x(transformed_point.x)
+    .y(transformed_point.y)
+    .z(transformed_point.z);
+}
+
+geometry_msgs::msg::Pose transformPose(
+  const geometry_msgs::msg::Pose & pose, const geometry_msgs::msg::TransformStamped & transform)
+{
+  geometry_msgs::msg::Pose transformed_pose;
+  tf2::doTransform(pose, transformed_pose, transform);
+
+  return transformed_pose;
+}
+
+geometry_msgs::msg::Pose transformPose(
+  const geometry_msgs::msg::Pose & pose, geometry_msgs::msg::Transform & transform)
+{
+  geometry_msgs::msg::TransformStamped transform_stamped;
+  transform_stamped.transform = transform;
+
+  return transformPose(pose, transform_stamped);
+}
+
+geometry_msgs::msg::Pose transformPose(
+  const geometry_msgs::msg::Pose & pose, const geometry_msgs::msg::Pose & pose_transform)
+{
+  tf2::Transform transform;
+  tf2::convert(pose_transform, transform);
+
+  geometry_msgs::msg::TransformStamped transform_msg;
+  transform_msg.transform = tf2::toMsg(transform);
+
+  return transformPose(pose, transform_msg);
+}
+
+// Transform pose in world coordinates to local coordinates
+/*
+geometry_msgs::msg::Pose inverseTransformPose(
+const geometry_msgs::msg::Pose & pose, const geometry_msgs::msg::TransformStamped & transform)
+{
+tf2::Transform tf;
+tf2::fromMsg(transform, tf);
+geometry_msgs::msg::TransformStamped transform_stamped;
+transform_stamped.transform = tf2::toMsg(tf.inverse());
+
+return transformPose(pose, transform_stamped);
+}
+*/
+
+// Transform pose in world coordinates to local coordinates
+geometry_msgs::msg::Pose inverseTransformPose(
+  const geometry_msgs::msg::Pose & pose, const geometry_msgs::msg::Transform & transform)
+{
+  tf2::Transform tf;
+  tf2::fromMsg(transform, tf);
+  geometry_msgs::msg::TransformStamped transform_stamped;
+  transform_stamped.transform = tf2::toMsg(tf.inverse());
+
+  return transformPose(pose, transform_stamped);
+}
+
+// Transform pose in world coordinates to local coordinates
+geometry_msgs::msg::Pose inverseTransformPose(
+  const geometry_msgs::msg::Pose & pose, const geometry_msgs::msg::Pose & transform_pose)
+{
+  tf2::Transform transform;
+  tf2::convert(transform_pose, transform);
+
+  return inverseTransformPose(pose, tf2::toMsg(transform));
+}
+
+// Transform point in world coordinates to local coordinates
+Eigen::Vector3d inverseTransformPoint(
+  const Eigen::Vector3d & point, const geometry_msgs::msg::Pose & pose)
+{
+  const Eigen::Quaterniond q(
+    pose.orientation.w, pose.orientation.x, pose.orientation.y, pose.orientation.z);
+  const Eigen::Matrix3d R = q.normalized().toRotationMatrix();
+
+  const Eigen::Vector3d local_origin(pose.position.x, pose.position.y, pose.position.z);
+  Eigen::Vector3d local_point = R.transpose() * point - R.transpose() * local_origin;
+
+  return local_point;
+}
+
+// Transform point in world coordinates to local coordinates
+geometry_msgs::msg::Point inverseTransformPoint(
+  const geometry_msgs::msg::Point & point, const geometry_msgs::msg::Pose & pose)
+{
+  const Eigen::Vector3d local_vec =
+    inverseTransformPoint(Eigen::Vector3d(point.x, point.y, point.z), pose);
+  geometry_msgs::msg::Point local_point;
+  local_point.x = local_vec.x();
+  local_point.y = local_vec.y();
+  local_point.z = local_vec.z();
+  return local_point;
+}
+
+double calcCurvature(
+  const geometry_msgs::msg::Point & p1, const geometry_msgs::msg::Point & p2,
+  const geometry_msgs::msg::Point & p3)
+{
+  // Calculation details are described in the following page
+  // https://en.wikipedia.org/wiki/Menger_curvature
+  const double denominator =
+    calcDistance2d(p1, p2) * calcDistance2d(p2, p3) * calcDistance2d(p3, p1);
+  if (std::fabs(denominator) < 1e-10) {
+    throw std::runtime_error("points are too close for curvature calculation.");
+  }
+  return 2.0 * ((p2.x - p1.x) * (p3.y - p1.y) - (p2.y - p1.y) * (p3.x - p1.x)) / denominator;
+}
+
+/**
+ * @brief Calculate offset pose. The offset values are defined in the local coordinate of the input
+ * pose.
+ */
+geometry_msgs::msg::Pose calcOffsetPose(
+  const geometry_msgs::msg::Pose & p, const double x, const double y, const double z)
+{
+  geometry_msgs::msg::Pose pose;
+  geometry_msgs::msg::Transform transform;
+  transform.translation = createTranslation(x, y, z);
+  transform.rotation = createQuaternion(0.0, 0.0, 0.0, 1.0);
+  tf2::Transform tf_pose;
+  tf2::Transform tf_offset;
+  tf2::fromMsg(transform, tf_offset);
+  tf2::fromMsg(p, tf_pose);
+  tf2::toMsg(tf_pose * tf_offset, pose);
+  return pose;
+}
+
+/**
+ * @brief Judge whether twist covariance is valid.
+ *
+ * @param twist_with_covariance source twist with covariance
+ * @return If all element of covariance is 0, return false.
+ */
+//
+bool isTwistCovarianceValid(const geometry_msgs::msg::TwistWithCovariance & twist_with_covariance)
+{
+  for (const auto & c : twist_with_covariance.covariance) {
+    if (c != 0.0) {
+      return true;
+    }
+  }
+  return false;
+}
+
+// NOTE: much faster than boost::geometry::intersects()
+std::optional<geometry_msgs::msg::Point> intersect(
+  const geometry_msgs::msg::Point & p1, const geometry_msgs::msg::Point & p2,
+  const geometry_msgs::msg::Point & p3, const geometry_msgs::msg::Point & p4)
+{
+  // calculate intersection point
+  const double det = (p1.x - p2.x) * (p4.y - p3.y) - (p4.x - p3.x) * (p1.y - p2.y);
+  if (det == 0.0) {
+    return std::nullopt;
+  }
+
+  const double t = ((p4.y - p3.y) * (p4.x - p2.x) + (p3.x - p4.x) * (p4.y - p2.y)) / det;
+  const double s = ((p2.y - p1.y) * (p4.x - p2.x) + (p1.x - p2.x) * (p4.y - p2.y)) / det;
+  if (t < 0 || 1 < t || s < 0 || 1 < s) {
+    return std::nullopt;
+  }
+
+  geometry_msgs::msg::Point intersect_point;
+  intersect_point.x = t * p1.x + (1.0 - t) * p2.x;
+  intersect_point.y = t * p1.y + (1.0 - t) * p2.y;
+  return intersect_point;
+}
+
+}  // namespace tier4_autoware_utils
