+++ control/autonomous_emergency_braking/src/node.cpp
@@ -19,6 +19,8 @@
 #include <tier4_autoware_utils/geometry/geometry.hpp>
 #include <tier4_autoware_utils/ros/marker_helper.hpp>
 
+#include <boost/geometry/strategies/agnostic/hull_graham_andrew.hpp>
+
 #include <pcl/filters/voxel_grid.h>
 #include <tf2/utils.h>
 
@@ -31,6 +33,8 @@
 #include <tf2_geometry_msgs/tf2_geometry_msgs.hpp>
 #endif
 
+#include <boost/geometry/algorithms/convex_hull.hpp>
+#include <boost/geometry/algorithms/within.hpp>
 namespace autoware::motion::control::autonomous_emergency_braking
 {
 using diagnostic_msgs::msg::DiagnosticStatus;
