+++ planning/behavior_path_planner/src/utils/avoidance/utils.cpp
@@ -28,6 +28,8 @@
 
 #include <tier4_planning_msgs/msg/avoidance_debug_factor.hpp>
 
+#include <boost/geometry/algorithms/correct.hpp>
+
 #include <lanelet2_routing/RoutingGraphContainer.h>
 
 #include <algorithm>
