+++ planning/behavior_path_planner/src/utils/path_safety_checker/safety_check.cpp
@@ -21,6 +21,8 @@
 #include "object_recognition_utils/predicted_path_utils.hpp"
 #include "tier4_autoware_utils/geometry/boost_polygon_utils.hpp"
 
+#include <boost/geometry/algorithms/overlaps.hpp>
+
 namespace behavior_path_planner::utils::path_safety_checker
 {
 void appendPointToPolygon(Polygon2d & polygon, const geometry_msgs::msg::Point & geom_point)
