+++ planning/obstacle_stop_planner/src/planner_utils.cpp
@@ -22,8 +22,10 @@
 #include <diagnostic_msgs/msg/key_value.hpp>
 
 #include <boost/format.hpp>
+#include <boost/geometry/algorithms/convex_hull.hpp>
 #include <boost/geometry/algorithms/distance.hpp>
 #include <boost/geometry/algorithms/within.hpp>
+#include <boost/geometry/strategies/agnostic/hull_graham_andrew.hpp>
 
 #include <pcl_conversions/pcl_conversions.h>
